nearby: sync from google3 at CL 548882463
Bug: 290645793
Test: presubmit
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:24f2d103d4a00405cf381c95f2fd99f5d11dab11)
Merged-In: I2ce4ee2d9b0df2b6561ad3b05fb3652051acb0bd

+++ apps/nearby/location/lbs/contexthub/nanoapps/nearby/filter_extension.cc
@@ -58,7 +58,7 @@ void FilterExtension::Update(
   }
   chreBleScanFilter scan_filter;
   scan_filter.rssiThreshold = CHRE_BLE_RSSI_THRESHOLD_NONE;
-  scan_filter.scanFilterCount = generic_filters->size();
+  scan_filter.scanFilterCount = static_cast<uint8_t>(generic_filters->size());
   scan_filter.scanFilters = generic_filters->data();
 
   chrexNearbyExtendedFilterConfig config;
